# openapi.yaml
openapi: 3.0.0
info:
  title: Service Envelope API v2.0
  version: 2.1.0
  description: REST API for the Service Envelope System. Includes OTP bearer
    authentication, Mongo-backed auth users and roles, generic email template
    fallbacks, schema v1.0.5 service validation, and unified delivery selection.
  contact:
    name: Service Envelope Team
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Development
  - url: https://api.mapua.edu.ph
    description: Production
paths:
...
  /api/OTP/send:
    post:
      tags:
        - Auth
      summary: Send OTP Login Code
      operationId: sendOtp
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpSendRequest"
      responses:
        "202":
          description: OTP send accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpSendResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          description: OTP was sent recently
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: OTP email could not be sent
  /api/OTP/verify:
    post:
      tags:
        - Auth
      summary: Verify OTP And Issue Bearer Token
      operationId: verifyOtp
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerifyRequest"
      responses:
        "200":
          description: OTP verified and bearer token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpVerifyResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "410":
          description: OTP expired
        "429":
          description: Too many OTP attempts
  /api/OTP/cancel:
    post:
      tags:
        - Auth
      summary: Cancel Active OTP Challenges
      operationId: cancelOtp
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpCancelRequest"
      responses:
        "200":
          description: Active OTP challenges cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpCancelResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/OTP/flush:
    post:
      tags:
        - Auth
      summary: Flush Stale OTP Challenges
      operationId: flushOtp
      security:
        - BearerAuth: []
      description: Requires a super_admin bearer token.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                consumedRetentionHours:
                  type: number
                  default: 24
      responses:
        "200":
          description: Stale OTP records flushed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpFlushResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/admin/email-templates/{templateId}:
    get:
      tags:
        - Admin
      summary: Get Email Template
      operationId: getEmailTemplate
      security: []
      parameters:
        - $ref: "#/components/parameters/TemplateIdParam"
      responses:
        "200":
          description: Template details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTemplate"
              examples:
                sample:
                  summary: Sample 200 response
                  value:
                    id: SERV-1-approval
                    name: Approval Request Template
                    subject: Approval needed for {{requestId}}
                    htmlBody: sample
                    description: sample
                    templateScope: generic
                    eventKey: sample
                    serviceType: sample
                    isActive: true
                    envelopeType: sample
                    phase: sample
                    createdAt: 2026-06-07T00:00:00.000Z
                    updatedAt: 2026-06-07T00:00:00.000Z
        "404":
          $ref: "#/components/responses/NotFound"
      x-codeSamples:
        - lang: bash
          label: cURL
          source: curl -X GET
            "http://localhost:8000/api/admin/email-templates/delivery-email-document"
    put:
      tags:
        - Admin
      summary: Update Email Template
      operationId: updateEmailTemplate
      security: []
      parameters:
        - $ref: "#/components/parameters/TemplateIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailTemplateUpdate"
            examples:
              sample:
                summary: Sample request
                value:
                  name: sample
                  subject: sample
                  htmlBody: sample
      responses:
        "200":
          description: Template updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
      x-codeSamples:
        - lang: bash
          label: cURL
          source: |-
            curl -X PUT "http://localhost:8000/api/admin/email-templates/delivery-email-document" \
              -H "Content-Type: application/json" \
              -d '{"name":"sample","subject":"sample","htmlBody":"sample"}'
    delete:
      tags:
        - Admin
      summary: Delete Email Template
      operationId: deleteEmailTemplate
      security: []
      parameters:
        - $ref: "#/components/parameters/TemplateIdParam"
      responses:
        "200":
          description: Template deleted
        "404":
          $ref: "#/components/responses/NotFound"
      x-codeSamples:
        - lang: bash
          label: cURL
          source: curl -X DELETE
            "http://localhost:8000/api/admin/email-templates/delivery-email-document"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    TemplateIdParam:
      name: templateId
      in: path
      required: true
      schema:
        type: string
      example: SERV-1-approval
      description: Email template ID.
...
  schemas:
    ApiRole:
      type: string
      enum:
        - super_admin
        - admin
        - requester
        - orchestrator
      example: requester
    AuthUserPublic:
      type: object
      properties:
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/ApiRole"
        name:
          type: string
...
    OtpSendRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
        purpose:
          type: string
          default: login
    OtpSendResponse:
      type: object
      properties:
        status:
          type: string
          example: sent
        expiresInSeconds:
          type: integer
          example: 600
        retryAfterSeconds:
          type: integer
          example: 60
    OtpVerifyRequest:
      type: object
      required:
        - email
        - code
      properties:
        email:
          type: string
          format: email
        code:
          type: string
          example: "123456"
        purpose:
          type: string
          default: login
    OtpVerifyResponse:
      type: object
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: string
          example: 1h
        user:
          $ref: "#/components/schemas/AuthUserPublic"
    OtpCancelRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
        purpose:
          type: string
          default: login
    OtpCancelResponse:
      type: object
      properties:
        status:
          type: string
          example: cancelled
        cancelledCount:
          type: integer
    OtpFlushResponse:
      type: object
      properties:
        status:
          type: string
          example: flushed
        deletedCount:
          type: integer
...
    EmailTemplate:
      type: object
      properties:
        id:
          type: string
          example: SERV-1-approval
        name:
          type: string
          example: Approval Request Template
        subject:
          type: string
          example: Approval needed for {{requestId}}
        htmlBody:
          type: string
        description:
          type: string
        templateScope:
          type: string
          enum:
            - generic
            - envelope
            - service
        eventKey:
          type: string
        serviceType:
          type: string
        isActive:
          type: boolean
        envelopeType:
          type: string
        phase:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    EmailTemplateUpdate:
      type: object
      required:
        - name
        - subject
        - htmlBody
      properties:
        name:
          type: string
        subject:
          type: string
        htmlBody:
          type: string
        description:
          type: string
        templateScope:
          type: string
          enum:
            - generic
            - envelope
            - service
        eventKey:
          type: string
        serviceType:
          type: string
        isActive:
          type: boolean
        envelopeType:
          type: string
        phase:
          type: string
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Request not found
        details:
          type: string
  responses:
    BadRequest:
      description: Bad Request - Invalid parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            sample:
              summary: Sample 400 response
              value:
                error: Request not found
                details: sample
    Unauthorized:
      description: Unauthorized - Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            sample:
              summary: Sample 401 response
              value:
                error: Request not found
                details: sample
    Forbidden:
      description: Forbidden - Authenticated user does not have the required role or ownership
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            insufficientRole:
              value:
                error: Insufficient permissions
    NotFound:
      description: Not Found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            sample:
              summary: Sample 404 response
              value:
                error: Request not found
                details: sample
...